Tester CSS-klasse på Fagord-logo under sidelast
