+++ vitest.config.ts
@@ -1,6 +1,8 @@
+import svgr from 'vite-plugin-svgr';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
+  plugins: [svgr()],
   resolve: {
     tsconfigPaths: true,
   },
